redo batch rendering
