+++ Artless/ALEngine/Include/ECS/Components/Transform.h
@@ -21,7 +21,7 @@ namespace ALEngine
 			***********************************************************************************/
 			struct Transform
 			{
-				Math::Vector2 position{ 0.0f, 0.0f };
+				Math::Vector3 position{ 0.0f, 0.0f, 0.0f };
 				Math::Vector2 scale{ 1.0f, 1.0f };
 				f32	rotation{ 0.0f };
 			};
