+++ Artless/ALEngine/Include/Graphics/stb_image.h
@@ -1025,14 +1025,14 @@ static int stbi__mad4sizes_valid(int a, int b, int c, int d, int add)
 static void* stbi__malloc_mad2(int a, int b, int add)
 {
     if (!stbi__mad2sizes_valid(a, b, add)) return NULL;
-    return stbi__malloc(a * b + add);
+    return stbi__malloc(static_cast<size_t>(a * b + add));
 }
 #endif
 
 static void* stbi__malloc_mad3(int a, int b, int c, int add)
 {
     if (!stbi__mad3sizes_valid(a, b, c, add)) return NULL;
-    return stbi__malloc(a * b * c + add);
+    return stbi__malloc(static_cast<size_t>(a * b * c + add));
 }
 
 #if !defined(STBI_NO_LINEAR) || !defined(STBI_NO_HDR) || !defined(STBI_NO_PNM)
