+++ Artless/bin/Assets/Shaders/mesh.frag
@@ -1,9 +1,9 @@
 #version 450 core
-layout (location = 0) out vec4 fColor;
+layout (location = 0) in vec3 fColor;
 
-uniform vec4 color = vec4(1.0f, 1.0f, 1.0f, 1.0f);
+layout (location = 0) out vec4 FragColor;
 
 void main()
 {
-	fColor =  color;
+    FragColor = vec4(fColor, 1.0);
 }
