+++ Artless/bin/Assets/Shaders/mesh.vert
@@ -1,10 +1,19 @@
 #version 450 core
 layout (location = 0) in vec2 aPos;
-layout (location = 1) in mat4 aInstanceModel;
+layout (location = 1) in vec2 aTexCoord;
+layout (location = 2) in vec3 aColor;
+layout (location = 3) in mat4 aInstanceMatrix;
 
-uniform mat4 model, view, proj;
+layout (location = 0) out vec3 vColor;
+layout (location = 1) out vec2 vTexCoord;
+
+uniform mat4 projection;
+uniform mat4 view;
 
 void main()
 {
-	gl_Position = proj * view * model * vec4(aPos, 0.0f, 1.0f);
+	vColor = aColor;
+	vTexCoord = aTexCoord;
+    gl_Position = projection * view * aInstanceMatrix * vec4(aPos, 0.0, 1.0);
+
 }
